+++ server/src/factories/utils/deleteFile.js
@@ -3,6 +3,10 @@ import fs from 'fs';
 
 export default link => {
   new Promise((resolve, reject) => {
+    if (!fs.existsSync(link)) {
+      return resolve();
+    }
+
     fs.unlink(link, err => {
       if (err) return reject(err);
       resolve();
